refactor(llm): move torchtrade/train -> torchtrade/llm/train
'train' read as if it owned all training; this is specifically the LLM GRPO trainer. Move it
under a torchtrade.llm package (LLM actors already live in torchtrade.actor; this mirrors the
examples/llm/train/ layout and leaves room to consolidate LLM code). Public import is now
`from torchtrade.llm.train import LLMTrainer`. Tests moved to tests/llm/train/ to match; all
torchtrade.train references updated across package, tests, example, and guide. History preserved
via git mv. Full suite: 1793 passed, 15 skipped.
