+++ torchtrade/llm/__init__.py
@@ -1,0 +1,1 @@
+"""LLM trading: actors live in torchtrade.actor; GRPO fine-tuning in torchtrade.llm.train."""
