+++ torchtrade/llm/train/__init__.py
@@ -1,4 +1,4 @@
 """Training utilities for TorchTrade LLM actors."""
-from torchtrade.train.llm_trainer import LLMTrainer
+from torchtrade.llm.train.llm_trainer import LLMTrainer
 
 __all__ = ["LLMTrainer"]
